Revert "fix: break testenv → externalmcp → access import cycle"
This reverts commit 1eea65ee3a1c88ccb0c5547c5823de405eb3ea86.
